requirements.txt
asgiref==3.2.7
astroid==2.4.2
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-model-utils==4.0.0
django-nested-admin==3.3.2
django-redis==4.12.1
django-redis-cache==2.1.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docx==0.2.4
docxtpl==0.10.0
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.1
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.25.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
install==1.3.1
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauth2client==4.1.3
pep8==1.7.1
Pillow==7.1.2
pkg-resources==0.0.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
PyJWT==1.7.1
pylint==2.5.3
python-docx==0.8.10
python-monkey-business==1.0.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requestes==0.0.1
requests==2.23.0
rsa==4.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
